Rename tuya base entity module
