+++ homeassistant/components/tuya/climate.py
@@ -24,8 +24,8 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import TuyaConfigEntry
-from .base import IntegerTypeData, TuyaEntity
 from .const import TUYA_DISCOVERY_NEW, DPCode, DPType
+from .entity import IntegerTypeData, TuyaEntity
 
 TUYA_HVAC_TO_HA = {
     "auto": HVACMode.HEAT_COOL,
