+++ homeassistant/components/tuya/entity.py
